refactor(ssa): decentralize compile-unit logic and drop lazyBuilder-by-unit
Five related cleanups of the compile-unit streaming pipeline, all
behavior-preserving (full scripts/ssa-test.sh: 79 packages pass):

1. Move per-language compile-unit partition + dependency extraction out
   of ssaapi/compile_units.go (which used hand-rolled regex) into each
   ast2ssa builder as ssa.Builder methods. Add a UnitPartitioner interface
   (PartitionCompileUnits / CompileUnitDependencies) to ssa.Builder with
   directory-based defaults on PreHandlerBase. go2ssa/java2ssa/python2ssa/
   c2ssa/ts2ssa now own their regex. CompileUnit/UnitRef types and shared
   path helpers move to ssa/compile_unit.go; ssaapi keeps only the
   engine-side topo/SCC ordering.

2. Drop the hardcoded languagePreHandlerBuildsFiles switch in ssaapi and
   use the existing builder.UsesDeferredFileBuild() method, so each AST
   decides whether it builds files in the pre-handler.

3. Extract the logging/recording helpers piled in ssa_compile_fs.go
   (heap/phase logging, compile-unit plan logging, file-performance
   recording, IR-save progress callback) into compile_units_log.go.

4. Remove the rebase-reintroduced lazyBuilder-by-unit tracking
   (lazyUnitMu / lazyBuildersByUnit / lazyBuilderUnitSet, BindLazyBuilderUnit,
   runWithCompileUnit, registerLazyBuilderForUnit, lazyBuildersForUnitSet,
   LazyBuildForUnits/lazyBuildForUnits, BuildForUnits, and the
   unitProvider/unitTaskObserver/unitTaskRunner machinery). The deferred
   build path (RunDeferredBuildsForUnits over the OrderedMap) plus the
   final LazyBuild() already cover function-body expansion; the per-unit
   lazy index was redundant. lazyTask reverts to a plain func().

5. Consolidate program.go's compile-unit lifecycle code into the new
   program_unit.go and merge program_unit_cleanup.go into it (release,
   memory-pressure, aggressive-clear, unit-key helpers).
