+++ common/yak/ssa/deferred_build_test.go
@@ -97,7 +97,7 @@ func TestRunDeferredBuildsForUnitsRestoresUnitDuringTask(t *testing.T) {
 	require.Equal(t, "", prog.CurrentCompileUnit())
 	require.Empty(t, ran)
 
-	prog.LazyBuildForUnits([]string{"unit-a"})
+	prog.LazyBuild()
 	require.Equal(t, []string{"nested"}, ran)
 }
 
@@ -140,134 +140,4 @@ func TestFinishAllowsLazyLibraryExpansion(t *testing.T) {
 		require.NoError(t, err)
 	default:
 	}
-}
-
-func TestLazyBuildForUnitsOnlyRunsMatchingLazyTasks(t *testing.T) {
-	prog := newDeferredBuildTestProgram(t, "lazy-build-unit-drain")
-	builder := prog.GetAndCreateFunctionBuilder("", string(MainFunctionName))
-
-	var ran []string
-	prog.BeginCompileUnit("unit-a")
-	builder.Function.AddLazyBuilder(func() {
-		ran = append(ran, "a")
-	})
-	prog.EndCompileUnit()
-
-	prog.BeginCompileUnit("unit-b")
-	builder.Function.AddLazyBuilder(func() {
-		ran = append(ran, "b")
-	})
-	prog.EndCompileUnit()
-
-	prog.LazyBuildForUnits([]string{"unit-a"})
-	require.Equal(t, []string{"a"}, ran)
-
-	prog.LazyBuild()
-	require.Equal(t, []string{"a", "b"}, ran)
-}
-
-func TestLazyBuildForUnitsKeepsBuilderOpenForLaterUnits(t *testing.T) {
-	prog := newDeferredBuildTestProgram(t, "lazy-build-unit-keeps-builder-open")
-	builder := prog.GetAndCreateFunctionBuilder("", string(MainFunctionName))
-
-	var ran []string
-	prog.BeginCompileUnit("unit-a")
-	builder.Function.AddLazyBuilder(func() {
-		ran = append(ran, "a")
-	})
-	prog.EndCompileUnit()
-
-	prog.LazyBuildForUnits([]string{"unit-a"})
-	require.Equal(t, []string{"a"}, ran)
-
-	prog.BeginCompileUnit("unit-b")
-	builder.Function.AddLazyBuilder(func() {
-		ran = append(ran, "b")
-	})
-	prog.EndCompileUnit()
-
-	prog.LazyBuildForUnits([]string{"unit-b"})
-	require.Equal(t, []string{"a", "b"}, ran)
-}
-
-func TestLazyBuildForUnitsClearsOnlyIndexedUnitBuilders(t *testing.T) {
-	prog := newDeferredBuildTestProgram(t, "lazy-build-unit-index-clears-unit")
-	builder := prog.GetAndCreateFunctionBuilder("", string(MainFunctionName))
-
-	prog.BeginCompileUnit("unit-a")
-	builder.Function.AddLazyBuilder(func() {})
-	prog.EndCompileUnit()
-
-	prog.BeginCompileUnit("unit-b")
-	builder.Function.AddLazyBuilder(func() {})
-	prog.EndCompileUnit()
-
-	require.Contains(t, prog.lazyBuildersByUnit, "unit-a")
-	require.Contains(t, prog.lazyBuildersByUnit, "unit-b")
-
-	prog.LazyBuildForUnits([]string{"unit-a"})
-
-	require.NotContains(t, prog.lazyBuildersByUnit, "unit-a")
-	require.Contains(t, prog.lazyBuildersByUnit, "unit-b")
-}
-
-func TestLazyBuildForUnitsDrainsIndexedBuildersRegisteredDuringBuild(t *testing.T) {
-	prog := newDeferredBuildTestProgram(t, "lazy-build-unit-index-drains-new-builders")
-	builder := prog.GetAndCreateFunctionBuilder("", string(MainFunctionName))
-
-	var ran []string
-	prog.BeginCompileUnit("unit-a")
-	builder.Function.AddLazyBuilder(func() {
-		require.Equal(t, "unit-a", prog.CurrentCompileUnit())
-		ran = append(ran, "first")
-		builder.Function.AddLazyBuilder(func() {
-			require.Equal(t, "unit-a", prog.CurrentCompileUnit())
-			ran = append(ran, "second")
-		})
-	})
-	prog.EndCompileUnit()
-
-	prog.LazyBuildForUnits([]string{"unit-a"})
-
-	require.Equal(t, "", prog.CurrentCompileUnit())
-	require.Equal(t, []string{"first", "second"}, ran)
-	require.NotContains(t, prog.lazyBuildersByUnit, "unit-a")
-}
-
-func TestLazyBuildForUnitsSkipsProgramCycles(t *testing.T) {
-	progA := newDeferredBuildTestProgram(t, "lazy-build-unit-cycle-a")
-	progB := newDeferredBuildTestProgram(t, "lazy-build-unit-cycle-b")
-	builderA := progA.GetAndCreateFunctionBuilder("", string(MainFunctionName))
-	builderB := progB.GetAndCreateFunctionBuilder("", string(MainFunctionName))
-
-	var ran []string
-	progA.BeginCompileUnit("unit-a")
-	builderA.Function.AddLazyBuilder(func() {
-		ran = append(ran, "a")
-	})
-	progA.EndCompileUnit()
-
-	progB.BeginCompileUnit("unit-a")
-	builderB.Function.AddLazyBuilder(func() {
-		ran = append(ran, "b")
-	})
-	progB.EndCompileUnit()
-
-	progA.AddUpStream(progB)
-	progB.AddUpStream(progA)
-
-	done := make(chan struct{})
-	go func() {
-		progA.LazyBuildForUnits([]string{"unit-a"})
-		close(done)
-	}()
-	select {
-	case <-done:
-	case <-time.After(2 * time.Second):
-		t.Fatal("lazy build for units did not finish across cyclic program graph")
-	}
-	require.ElementsMatch(t, []string{"a", "b"}, ran)
-
-	progA.LazyBuildForUnits([]string{"unit-a"})
-	require.ElementsMatch(t, []string{"a", "b"}, ran)
 }
