+++ common/yak/ssa/func.go
@@ -37,7 +37,6 @@ func (p *Program) NewFunctionWithParent(name string, parent *Function) *Function
 		SideEffects: make([]*FunctionSideEffect, 0),
 		builder:     nil,
 	}
-	p.BindLazyBuilderUnit(f.LazyBuilder)
 	f.SetName(name)
 	f.SetProgram(p)
 	p.SetVirtualRegister(f)
